+++ src/bank/dao/CurrencyDAO.java
@@ -42,7 +42,7 @@ public class CurrencyDAO implements DAO<Currency> {
     */
    @SuppressWarnings("unchecked")
    @Override
-   public Optional<Currency> findById(Integer id) throws DAOException {
+   public Optional<Currency> findById(Long id) throws DAOException {
       try {
          return (Optional<Currency>) database.findById(Currency.class, id);
       } catch (DatabaseException e) {
@@ -115,7 +115,7 @@ public class CurrencyDAO implements DAO<Currency> {
     * @throws DAOException If an error occurs while updating the entity.
     */
    @Override
-   public void update(Integer id, Currency entity) throws DAOException {
+   public void update(Long id, Currency entity) throws DAOException {
       try {
          database.update(Currency.class, id, entity);
       } catch (DatabaseException e) {
@@ -130,7 +130,7 @@ public class CurrencyDAO implements DAO<Currency> {
     * @throws DAOException If an error occurs while deleting the entity.
     */
    @Override
-   public void delete(Integer id) throws DAOException {
+   public void delete(Long id) throws DAOException {
       try {
          database.delete(Currency.class, id);
       } catch (DatabaseException e) {
